test(release-tools): move FakeGitHubApi to tests, add PHP matrix + coverage
