+++ .github/workflows/test-release-tools.yml
@@ -15,6 +15,11 @@ permissions:
 jobs:
   test:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        php-version: ['8.3', '8.4']
+    name: test (php ${{ matrix.php-version }})
     defaults:
       run:
         working-directory: tools/release
@@ -26,11 +31,21 @@ jobs:
       - name: Set up PHP
         uses: shivammathur/setup-php@7c071dfe9dc99bdf297fa79cb49ea005b9fcadbc # v2.37.1
         with:
-          php-version: '8.3'
+          php-version: ${{ matrix.php-version }}
+          coverage: pcov
           tools: composer
 
       - name: Install dependencies
         run: composer install --no-interaction --no-progress
 
-      - name: PHPUnit
-        run: composer test
+      - name: PHPUnit with coverage
+        run: vendor/bin/phpunit --coverage-text --coverage-clover coverage.xml
+
+      - name: Upload coverage to Codecov
+        if: matrix.php-version == '8.3'
+        uses: codecov/codecov-action@v5
+        with:
+          files: tools/release/coverage.xml
+          flags: release-tools
+          token: ${{ secrets.CODECOV_TOKEN }}
+          fail_ci_if_error: false
