+++ tools/release/.gitignore
@@ -1,2 +1,3 @@
 /vendor/
 /.phpunit.result.cache
+/coverage.xml
